Refactor directory structure of components to accomodate sharing of components between Marketplace and Collection
